+++ test/scope_blue_print_test.dart
@@ -12,10 +12,11 @@ void main() {
     group('example', () {
       test('should provide a blue print with to nodes and one dependency', () {
         final scopeBluePrint = ScopeBluePrint.example();
-        final dependency = scopeBluePrint.nodeOverrides.first;
-        final subScope = scopeBluePrint.scopeOverrides.first;
-        final node = subScope.nodeOverrides.first as NodeBluePrint<int>;
-        final customer = subScope.nodeOverrides.last as NodeBluePrint<int>;
+        final dependency = scopeBluePrint.nodesFromConstructor.first;
+        final subScope = scopeBluePrint.childrenFromConstructor.first;
+        final node = subScope.nodesFromConstructor.first as NodeBluePrint<int>;
+        final customer =
+            subScope.nodesFromConstructor.last as NodeBluePrint<int>;
         expect(scopeBluePrint.toString(), scopeBluePrint.key);
         expect(dependency.key, 'dependency');
         expect(node.key, 'node');
@@ -50,8 +51,8 @@ void main() {
         expect(scopeBluePrint.key, 'a');
 
         void expectNode(int i, String key, dynamic value) {
-          expect(scopeBluePrint.nodeOverrides[i].key, key);
-          expect(scopeBluePrint.nodeOverrides[i].initialProduct, value);
+          expect(scopeBluePrint.nodesFromConstructor[i].key, key);
+          expect(scopeBluePrint.nodesFromConstructor[i].initialProduct, value);
         }
 
         expectNode(0, 'int', 5);
@@ -60,24 +61,29 @@ void main() {
         expectNode(3, 'bool', true);
         expectNode(4, 'bluePrint', 8);
 
-        expect(scopeBluePrint.scopeOverrides.length, 2);
-        expect(scopeBluePrint.scopeOverrides.first.key, 'b');
-        expect(scopeBluePrint.scopeOverrides.first.scopeOverrides.length, 1);
+        expect(scopeBluePrint.childrenFromConstructor.length, 2);
+        expect(scopeBluePrint.childrenFromConstructor.first.key, 'b');
         expect(
-          scopeBluePrint.scopeOverrides.first.scopeOverrides.first.key,
+          scopeBluePrint
+              .childrenFromConstructor.first.childrenFromConstructor.length,
+          1,
+        );
+        expect(
+          scopeBluePrint
+              .childrenFromConstructor.first.childrenFromConstructor.first.key,
           'c',
         );
-        expect(scopeBluePrint.scopeOverrides.last.key, 'c');
+        expect(scopeBluePrint.childrenFromConstructor.last.key, 'c');
 
         expect(
-          scopeBluePrint.scopeOverrides.first.scopeOverrides.first.nodeOverrides
-              .first.key,
+          scopeBluePrint.childrenFromConstructor.first.childrenFromConstructor
+              .first.nodesFromConstructor.first.key,
           'x',
         );
 
         expect(
-          scopeBluePrint.scopeOverrides.first.scopeOverrides.first.nodeOverrides
-              .first.initialProduct,
+          scopeBluePrint.childrenFromConstructor.first.childrenFromConstructor
+              .first.nodesFromConstructor.first.initialProduct,
           123,
         );
       });
@@ -151,7 +157,7 @@ void main() {
           // which overrides the with key nodeConstructedByChildScope
           const overridenScope = ScopeBluePrint(
             key: 'childScopeConstructedByParent',
-            nodeOverrides: [
+            nodes: [
               NodeBluePrint<int>(
                 key: 'nodeConstructedByChildScope',
                 initialProduct: 6,
@@ -162,7 +168,7 @@ void main() {
           // Instantiate ExampleScopeBluePrint
           // and override the childScopeConstructedByParent
           final bluePrint = ExampleScopeBluePrint(
-            scopeOverrides: [
+            childrenFromConstructor: [
               overridenScope,
             ],
           );
@@ -220,7 +226,7 @@ void main() {
               .writeImageFile('test.graphs.example_scope_blue_print.dot');
         });
 
-        test('and apply nodeOverrides when provided', () {
+        test('and apply nodesFromConstructor when provided', () {
           const replacedBluePrint = NodeBluePrint<int>(
             key: 'nodeBuiltByParent',
             initialProduct: 111,
@@ -228,7 +234,7 @@ void main() {
 
           final rootScope = Scope.example();
           final scope = ExampleScopeBluePrint(
-            nodeOverrides: [replacedBluePrint],
+            nodes: [replacedBluePrint],
           ).instantiate(
             scope: rootScope,
           );
@@ -244,7 +250,7 @@ void main() {
         test('when the keys are the same', () {
           const bluePrint = ScopeBluePrint(
             key: 'root',
-            nodeOverrides: [
+            nodes: [
               NodeBluePrint<int>(key: 'node', initialProduct: 5),
               NodeBluePrint<int>(key: 'node', initialProduct: 6),
               NodeBluePrint<int>(key: 'node1', initialProduct: 5),
@@ -288,13 +294,15 @@ void main() {
       });
 
       test('should return the node with the given key', () {
-        final bluePrint = ScopeBluePrint.example().scopeOverrides.first;
+        final bluePrint =
+            ScopeBluePrint.example().childrenFromConstructor.first;
         final node = bluePrint.findNode<int>('node');
         expect(node, isNotNull);
       });
 
       test('should throw if the type does not match', () {
-        final bluePrint = ScopeBluePrint.example().scopeOverrides.first;
+        final bluePrint =
+            ScopeBluePrint.example().childrenFromConstructor.first;
 
         expect(
           () => bluePrint.findNode<String>('node'),
@@ -320,18 +328,22 @@ void main() {
         test('with the given nodes', () {
           final bluePrint = ScopeBluePrint.example();
           final copy = bluePrint.copyWith(modifiedNodes: []);
-          expect(copy.nodeOverrides, bluePrint.nodeOverrides);
+          expect(copy.nodesFromConstructor, bluePrint.nodesFromConstructor);
         });
 
         test('with the given subScopes', () {
           final bluePrint = ScopeBluePrint.example();
           final otherSubScopes = <ScopeBluePrint>[];
           final copy = bluePrint.copyWith(modifiedScopes: otherSubScopes);
-          expect(copy.scopeOverrides, same(bluePrint.scopeOverrides));
+          expect(
+            copy.childrenFromConstructor,
+            same(bluePrint.childrenFromConstructor),
+          );
         });
 
         test('with the given overrides', () {
-          final bluePrint = ScopeBluePrint.example().scopeOverrides.first;
+          final bluePrint =
+              ScopeBluePrint.example().childrenFromConstructor.first;
           const overriddenNode = NodeBluePrint<int>(
             key: 'node',
             initialProduct: 5,